Update PinkPill.pm
Changes default value of 'src_folders' and 'inc_folders' to 'code'
Added function `folder_file` to spilt a full path into a folder and the file in that folder. This is not a member function.
Added function `compilation_required` that will work out if a source file needs to be recompiled or not. Currently will always say compilation is required.
Added function `compile` that will handle compiling one file, it will not verify if compilation is actually required, this has moved most of the logic for actually compiling a file out of `compile_files`

These changes where made via github web gui and as such are untested!

updated version to 0.1.0